requirements.txt
pillow==12.1.1
pydantic==2.12.5
PyYAML==6.0.3
ruff==0.15.2
